Ver 0.a.d 20240819 	1. 사용하지 않은 GPIO 셋팅을 Output Low 로 설정 	2. 모자이크 동작을 할 경우 위아래 싱크가 맞지 않는 문제 보와 	 - 모자이크가 동작하면 상단에 영상 프레임이 빠지는 문제 확인 	 - 위아래 영상에 프레임을 실시간으로 확인하면서 두 영상에 프레임이 10프레임(약 0.5초)이상 차이 날 경우 	  두 프레임이 같아질 때 까지 I프레임을 제외한 다른 프레임은 버림 	 - 버리는 중간에 맞춰 질 경우 다음 I 프레임이 올때까진 다 버림 	3. 택배 캡쳐 영상 촬영 중 실시간 스트리밍이 들어오면 스트리밍도 열고 택배 영상도 보내서 SPI 혼선 	 - 스트리밍 중에는 택배 보내는 내용을 삭제

+++ setup.sh
@@ -1,11 +1,34 @@
 # F/W Update
-rm /tmp/mnt/sdcard/isc_bak
-mv /tmp/mnt/sdcard/isc /tmp/mnt/sdcard/isc_bak
-cp /dev/shm/isc /tmp/mnt/sdcard/isc
-chmod 777 /tmp/mnt/sdcard/isc
-
+if [ -e "/dev/shm/system.bin" ]; then
+	echo "F/W System Change"
+	flashcp /dev/shm/system.bin /dev/mtd5
+	echo "Flashcp Finish!!"
+else
+	echo "F/W Card Change"
+	rm /tmp/mnt/sdcard/isc_bak
+	mv /tmp/mnt/sdcard/isc /tmp/mnt/sdcard/isc_bak
+	cp /dev/shm/isc /tmp/mnt/sdcard/isc
+	chmod 777 /tmp/mnt/sdcard/isc
+fi
 
 # Tag Update
-flashcp /dev/shm/tag.bin /dev/mtd1
+if [ -e "/dev/shm/tag.bin" ]; then
+	echo "tag change start!"
+	flashcp /dev/shm/tag.bin /dev/mtd1
+	echo "Flashcp Finish!!!"
+fi
 
+# Executation Update
+if [ -e "/dev/shm/vexe.sh" ]; then
+	echo "vexe.sh chnge start!"
+	cp /dev/shm/vexe.sh /tmp/mnt/sdcard
+fi
+
+# Mute Update
+if [ -e "/dev/shm/mute_msg.sh" ]; then
+	echo "Mute chnge start!"
+	cp /dev/shm/mute_msg.sh /tmp/mnt/sdcard
+fi
+
+sync
 sync
